reference-app/trial/requirements.dev.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile --output-file=requirements.dev.txt requirements/develop.in
#
astroid==2.8.4
    # via pylint
atomicwrites==1.4.0
    # via pytest
attrs==21.2.0
    # via pytest
black==21.10b0
    # via -r requirements/develop.in
certifi==2021.10.8
    # via requests
charset-normalizer==2.0.7
    # via requests
click==8.0.3
    # via
    #   black
    #   flask
colorama==0.4.4
    # via
    #   click
    #   pylint
    #   pytest
deprecated==1.2.13
    # via opentelemetry-api
flake8==4.0.1
    # via -r requirements/develop.in
flask==2
    # via -r requirements\common.in
googleapis-common-protos==1.53.0
    # via opentelemetry-exporter-jaeger-proto-grpc
grpcio==1.41.1
    # via opentelemetry-exporter-jaeger-proto-grpc
idna==3.3
    # via requests
iniconfig==1.1.1
    # via pytest
isort==5.10.0
    # via pylint
itsdangerous==2.0.1
    # via flask
jaeger-client==4.8.0
    # via -r requirements\common.in
jinja2==3.0.2
    # via flask
lazy-object-proxy==1.6.0
    # via astroid
markupsafe==2.0.1
    # via jinja2
mccabe==0.6.1
    # via
    #   flake8
    #   pylint
mypy-extensions==0.4.3
    # via black
opentelemetry-api==1.6.2
    # via
    #   -r requirements\common.in
    #   opentelemetry-exporter-jaeger-proto-grpc
    #   opentelemetry-exporter-jaeger-thrift
    #   opentelemetry-instrumentation
    #   opentelemetry-instrumentation-flask
    #   opentelemetry-instrumentation-requests
    #   opentelemetry-instrumentation-wsgi
    #   opentelemetry-sdk
opentelemetry-exporter-jaeger==1.6.2
    # via -r requirements\common.in
opentelemetry-exporter-jaeger-proto-grpc==1.6.2
    # via opentelemetry-exporter-jaeger
opentelemetry-exporter-jaeger-thrift==1.6.2
    # via opentelemetry-exporter-jaeger
opentelemetry-instrumentation==0.25b2
    # via
    #   -r requirements\common.in
    #   opentelemetry-instrumentation-flask
    #   opentelemetry-instrumentation-requests
    #   opentelemetry-instrumentation-wsgi
opentelemetry-instrumentation-flask==0.25b2
    # via -r requirements\common.in
opentelemetry-instrumentation-requests==0.25b2
    # via -r requirements\common.in
opentelemetry-instrumentation-wsgi==0.25b2
    # via opentelemetry-instrumentation-flask
opentelemetry-sdk==1.6.2
    # via
    #   -r requirements\common.in
    #   opentelemetry-exporter-jaeger-proto-grpc
    #   opentelemetry-exporter-jaeger-thrift
opentelemetry-semantic-conventions==0.25b2
    # via
    #   opentelemetry-instrumentation-flask
    #   opentelemetry-instrumentation-requests
    #   opentelemetry-instrumentation-wsgi
    #   opentelemetry-sdk
opentelemetry-util-http==0.25b2
    # via
    #   opentelemetry-instrumentation-flask
    #   opentelemetry-instrumentation-requests
    #   opentelemetry-instrumentation-wsgi
opentracing==2.4.0
    # via jaeger-client
packaging==21.2
    # via pytest
pathspec==0.9.0
    # via black
platformdirs==2.4.0
    # via
    #   black
    #   pylint
pluggy==1.0.0
    # via pytest
protobuf==3.19.1
    # via googleapis-common-protos
py==1.10.0
    # via pytest
pycodestyle==2.8.0
    # via flake8
pyflakes==2.4.0
    # via flake8
pylint==2.11.1
    # via -r requirements/develop.in
pyparsing==2.4.7
    # via packaging
pytest==6.2.5
    # via -r requirements/develop.in
regex==2021.11.2
    # via black
requests==2.26.0
    # via -r requirements\common.in
six==1.16.0
    # via
    #   grpcio
    #   thrift
threadloop==1.0.2
    # via jaeger-client
thrift==0.15.0
    # via
    #   jaeger-client
    #   opentelemetry-exporter-jaeger-thrift
toml==0.10.2
    # via
    #   pylint
    #   pytest
tomli==1.2.2
    # via black
tornado==6.1
    # via
    #   jaeger-client
    #   threadloop
typing-extensions==3.10.0.2
    # via
    #   astroid
    #   black
    #   pylint
urllib3==1.26.7
    # via requests
werkzeug==2.0.2
    # via flask
wrapt==1.13.3
    # via
    #   astroid
    #   deprecated
    #   opentelemetry-instrumentation

# The following packages are considered to be unsafe in a requirements file:
# setuptools
